Add user-defined spans to GPU timelines in activity traces
Summary: Updated ActivityProfiler so that user defined spans can be shown on the GPU timelines. This is shown on the specific GPU stream that the user span is correlated with (denoted as "stream <resourceId> user"). So a single span on the CPU side could be shown on multiple GPU streams.

Reviewed By: gdankel

Differential Revision: D25074836

fbshipit-source-id: e7c1a2b2942f1921a267ab64e0332be363dccd52