Change to safe thread name implementation
Summary:
Passing an invalid thread handle to pthread_getname_np may crash the process. This can happen when we try to get the name for a thread that has exited.
Switch to a robust implementation - which unfortunately is linux specific.

We can re-evaluate this later - another approach is to get the thread names from within the threads themselves as they are logging events.

Reviewed By: chaekit

Differential Revision: D26857000

fbshipit-source-id: ad41d8190e1966965468213900b48e1cee087d5e