Change to safe thread name implementation
Summary:
Passing an invalid thread handle to pthread_getname_np may crash the process. This can happen when we try to get the name for a thread that has exited.

Switch to a robust implementation - which unfortunately is linux specific.
We can re-evaluate this later - another approach is to get the thread names from within the threads themselves as they are logging events.

Differential Revision: D27003592

fbshipit-source-id: b4d4bcd34f8a6d61e8d12581c91c39d8d9a18d40