Log CUPTI errors on libkineto initialization
Summary: When libkineto is initialized from the PyTorch Profiler, if it fails we will not know why because errors are not reported. Reporting errors is not always safe, e.g. if init happens from static initialization or a dlopen library constructor function, so add a flag to specify whether to log.

Differential Revision: D27003561

fbshipit-source-id: c2e4ffcde5a4993d6044cbab03e6cd41d5af048f