Log CUPTI errors on libkineto initialization (#119)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/119

When libkineto is initialized from the PyTorch Profiler, if it fails we will not know why because errors are not reported. Reporting errors is not always safe, e.g. if init happens from static initialization or a dlopen library constructor function, so add a flag to specify whether to log.

Reviewed By: chaekit

Differential Revision: D27003561

fbshipit-source-id: bb4768f1940d4183de37bf007d547207a0da3bea