Change to safe thread name implementation (#117)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/117

Passing an invalid thread handle to pthread_getname_np may crash the process. This can happen when we try to get the name for a thread that has exited.

Switch to a robust implementation - which unfortunately is linux specific.
We can re-evaluate this later - another approach is to get the thread names from within the threads themselves as they are logging events.

Reviewed By: chaekit

Differential Revision: D27003592

fbshipit-source-id: 16c34c4a3b0cbf5686c3414995c7ca44227f4afb