Fix cbid name lookup bounds check and remove static assert (#121)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/121

The static assert adds additional work for PyTorch folks when updating the PyTorch Cuda version.

Note: It was added to notify us when the table needs updating, which is when PyTorch adds support for a new version of Cuda.
But if not using this approach we need another automatic reminder to add the support.

Also, the bounds check should use the names array size not the enum size, as otherwise it will break when running a version compiled for one version of Cuda on a newer version.

Reviewed By: chaekit

Differential Revision: D27084725

fbshipit-source-id: 36a8d61da5979f9e6c98832cd8abffd426b28673
(cherry picked from commit 868cf8117b605ba3253ce079016a971b47178036)
Signed-off-by: Eli Uriegas <eliuriegas@fb.com>