add JSON-formatted metadata for ClientTraceActivity (#164)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/164

Plan is to deprecate ClientTraceActivity and use a single concrete data structure to log all client side activities :`GenericTraceActivity`. CTA has a lot of explicit fields for metadata that are rarely used which results in sparse definition of activities. Instead of adding a new field, encode the key/value as a string. Note, more type-safe and rich alternative is `std::map` but we are not using it for performance reason.

Reviewed By: gdankel

Differential Revision: D27298829

fbshipit-source-id: b7a82081f983c5502f6d53104fc2d01ddee77b08