Rename theoretical occupancy to est. achieved occupancy (#210)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/210

Theoretical occupancy is used in Nvidia docs to mean the occupancy a kernel could achieve given sufficient input parallelism (grid size).

It's more interesting in a trace to know what occupancy is actually achieved. Then if it's low, theoretical occupancy can be studied for each kernel to estimate improvement from increasing grid size. But that's something we can do as part of a recommendation or tutorial.

Reviewed By: ilia-cher

Differential Revision: D28327605

fbshipit-source-id: 80b54a955fa2885dbbd45006c45f2fb0d9fca2d2