tb_plugin: fix startup hang when no profile data
When starting up tensorboard and the plugin in a directory where there
is no profile data, `TorchProfilerPlugin.is_active()` hangs forever. This
makes the tensorboard server hang, and client-side webpage remain blank.

This commit fixes the issue by ensuring that the thread monitoring the
run directories for data notifies the `_is_active_initialized_event`
flag not only when data is found, but also when the search is complete
and no data was found.

A simple unit test is added which validates the fix: without this
commit, the test hangs indefinitely.