Read base config before event profiler starts
Summary:
CUPTI has a known deadlock issue when programming h/w counters via the libkineto EventProfiler.
The workaround is to only program counters immediately upon context creation. A previous config change caused the programming to be delayed somewhat due to a delay in reading the configuration file.

This change forces the config file to be read sooner to avoid the deadlock issue while Nvidia is working on back porting the fix to Cuda11.0.

Reviewed By: xw285cornell

Differential Revision: D29955172

fbshipit-source-id: 2277a3787d8c1e508526f07e5a94d7619a61d844