Use deque for trace activities to avoid large vector resize
Summary: D31912779 discovered an perf issue of the observer that when we push new activities, we potentiall will resize the vector, and eveytime, we have to reallocate. This can be observed by doubling gaps in the trace for no obvious reason. The fix in D31912779 is to try to reserve enough buffer so that we don't resize. But I'm not sure how this can be enforced. This diff tries a simpler alternative: using std::deque instead std::vector for actitives push back. We will have to grow the deque but since the deque has fixed sized chunks, we can bound the overhead and have some consistency in the cost.

Reviewed By: gdankel

Differential Revision: D32339265

fbshipit-source-id: 14c13e7f60c1e4d85468194dc3d407c3c51843a4