Add Callback API centralized handler (#456)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/456

# Summary
CUPTI only permits one function to be registered to handle callbacks. In this change we add a level or indirection for callbacks
thus enabling any profilers in Kineto to access cupti callback interface.

PS: added some fixes for the case when HAS_CUPTI = false

## Use cases
1. Tracking CUDA context create/destroy for initializing the EventProfiler. Migrated this to use the new CallbackAPI module.
2. For Cupti Range profiler we can use callbacks for logging kernel names.

In the future we can also envision a Callback profiler to track copies/kernel arguments etc.

## Design decisions
1. One design choice we made is to only support simple function pointers in order to speed up the implementation for fast path.
2. __cupti_callback_exchange() function does the indirection. To keep it really lightweight it uses jump tables kind of structure making it a bit verbose but performant.

Reviewed By: gdankel

Differential Revision: D31704969

fbshipit-source-id: e2582953d13cd1b8092f3071abd597a87bf3e9af