Enable CUPTI Profiler API to track multiple threads/context (#544)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/544

# Summary

The CUPTI Ranger profiler API is confined to each CUDA context. This makes it harder to use in applications that are multi-threaded, such as pytorch models
with different threads for forward and backward pass computation.

In this patch we leverage CUPTI Callback API to track the CUDA contexts and kernel launches.
* Also, introduce an asynchronous mode to start and stop the CUPTI Range profiling API that is triggered using Kernel launch callbacks
* Track the name of kernels to use with auto-range model as CUPTI only provides a number for range ID.

## Testing
1. Enabled mocking the callbacks for Cuda context create/destroy and kernel launches.
2. Added a unit test to check callback triggered logic and asynchronous start and stop of ranges.

Reviewed By: aaronenyeshi

Differential Revision: D32008907

fbshipit-source-id: 9bce7431b090f2a999a3d8127c9ad789c253f004