Change output_membuffer to copy and own GenericTraceActivity objects (#553)
Summary:
This is a fix for a break at https://github.com/pytorch/kineto/pull/464
This change was adding pointers to temporaries (stack addresses) to the list of output events.  Specifically it was taking a const &, assuming it was referring to a heap object, deciding it owned it now, and ignoring any life-cycle concerns.

This change mirrors the handling used for GpuActivity.
1) change ActivityLogger::handleGenericActivity() to actually take a GenericTraceActivity instead of the abstract base.
2) concrete GenericTraceActivity can now be used with make_unique<>. Use addActivityWrapper (as GpuActivity does), which copies the object and maintains lifecycle via wrappers_
3) CLEANUP: added a ActivityLogger::handleActivity(const ITraceActivity) in case someone actually wants to log an abstract base event (at their peril).

Pull Request resolved: https://github.com/pytorch/kineto/pull/553

Test Plan: Tested in Kineto CI, by AMD in PyTorch Profiler, and internal CI.

Reviewed By: chaekit

Differential Revision: D34648204

Pulled By: aaronenyeshi

fbshipit-source-id: 5e6a8c6557abf0ecb426a657e1968795f3efeb38