Reset config state on remote Configs (#554)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/554

On demand configs will have options that need to be set from scratch. the Config has logic to set default options etc.
We need to reset the entire state so on-demand configs get parsed correctly.

This was causing a bug where if PROFILE_START_TIME is not present, instead of setting default, the 2nd Config onwards would get stale start times.

Reviewed By: chaekit

Differential Revision: D34661638

fbshipit-source-id: afe98fbf9d617103a279b4d6581ff1e5c5f09360