Update HAS_CUPTI_RANGE_PROFILER to CUDA 11+ for CUPTI Range Profiler APIs and remove init-statement in seletion statement
Summary:
The CuptiRangeProfilerAPI is using APIs from CUPTI in CUDA 11 such as CUpti_Profiler_GetCounterAvailability_Params and cuptiProfilerGetCounterAvailability. These are not available in CUDA 10.2 which caused PyTorch CI to crash.

Also, move the init statement outside of the if statement, which is a C++17 feature, but PyTorch uses C++14.

Differential Revision: D34690168

Pulled By: aaronenyeshi

fbshipit-source-id: 1bfcf3df2b03b6d7912db57c8bb6a81d2bccc78c