Update HAS_CUPTI_RANGE_PROFILER to CUDA 11+ for CUPTI Range Profiler APIs and remove init-statement in selection statement (#555)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/555

The CuptiRangeProfilerAPI is using APIs from CUPTI in CUDA 11 such as CUpti_Profiler_GetCounterAvailability_Params and cuptiProfilerGetCounterAvailability. These are not available in CUDA 10.2 which caused PyTorch CI to crash.

Also, move the init statement outside of the if statement, which is a C++17 feature, but PyTorch uses C++14.

Test Plan: CI Tests and PyTorch build in C++14.

Reviewed By: malfet

Differential Revision: D34690168

Pulled By: aaronenyeshi

fbshipit-source-id: 619570452130aabe29d83e12459c366b99a7f830