Add fixes for pytorch build CI (#575)
Summary:
Adding fixes for pytorch CI. Overall this add new files for CUPTI Range Profiler but disables out some functionality.

## Fixe also include
1. Update structure initialization to be complete, fails on gcc compilers
2. Remove unnecessary header <unistd> that fails on windows
3. Add USE_CUPTI_RANGE_PROFILER define to prevent some code from being included

In the future, we will need to be able to link nvperf_host.so in the build , this is required for range profiler to work.

Pull Request resolved: https://github.com/pytorch/kineto/pull/575

Test Plan:
See CI results in Linked pytoch PR
Looks good on platform combinations
https://github.com/pytorch/pytorch/pull/75206

Reviewed By: aaronenyeshi

Differential Revision: D35366236

Pulled By: briancoutinho

fbshipit-source-id: 900eb2a4accad753e4440ef72a8ae4c888cdf691