guard against CUPTI availability at runtime
Summary:
we only have `HAS_CUPTI` macro which guards against header availability. the issue is devvm now(or has it always been?) ships with cuda headers meaning this will always be true for internal fleet(not for OSS though).
so all the CUDA/CUPTI calls will just fail with warning due to missing cupti.so and cuda.so.

run the cudaGetDeviceCount which still returns 0 on missing symbol as a heuristic to detect GPU machine

Reviewed By: aaronenyeshi

Differential Revision: D35796924

fbshipit-source-id: 8cb4fa658c20643480aa5c873ff4e75636a6f174