Add methods to bridge the transition to std::unique_ptr<GenericTraceActivity> (#616)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/616

The goal is to replace:
```
struct CpuTraceBuffer {
  TraceSpan span{0, 0, "none"};
  int gpuOpCount;
  std::deque<GenericTraceActivity> activities;
};
```
with
```
struct CpuTraceBuffer {
  TraceSpan span{0, 0, "none"};
  int gpuOpCount;
  std::deque<std::unique_ptr<GenericTraceActivity>> activities;
};
```
This will add a lot of safety to Kineto (which currently uses a lot of raw pointers to elements in containers) and allow clients to keep track of which event is which even if `activities` is reordered.

However, this is a breaking change and will break clients who are  expecting the old behavior. To bridge the change, this PR introduces two methods to CpuTraceBuffer: An emplace method which can be changed to call std::make_unique, and a toRef method which will dereference the pointer when applicable. Once all callers use these methods, we can change `activities` without fear.

Reviewed By: aaronenyeshi

Differential Revision: D36679292

fbshipit-source-id: f71b921e65475b70628e46ae473badc9a4cc2e0c