Don't wrap GenericTraceActivities in CuptiRangeProfiler and CuptiActivityProfiler (#625)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/625

Now that CpuTraceBuffer stores its activities as unique_ptr we can safely use the unsafe `handleActivity` method without having to worry about the internal storage details of the container.

mwootton Is there an analogous change that we need to make for RocTracer?

To give some context: I agree with the assessment in https://github.com/pytorch/kineto/pull/553 that lifetime and ownership is not robustly handled in Kineto's internals. From the PyTorch profiler's perspective we want to be able to give Kineto an event and later be able to keep track of which events we provided and which events Kineto added. There isn't a concept of an ID in kineto events so instead I elected to use the address and just make the address stable.

Reviewed By: aaronenyeshi

Differential Revision: D37406957

fbshipit-source-id: 809e59a08dd6fb50361a1f577db9d7f8017dd938