Correlate cudaGraphLaunch to the first kernel
Summary: There was no correlation flow from `cudaGraphLaunch` to kernels which affected readability. This diff adds the flow correctly. The only caveat is that it only points to the first kernel launch and not the subsequent ones, as this is a limitation to Chrome tracing(does not support multiple flows). Since `cudaGraph`-launched kernels are grouped next to one another, this should be sufficient to improve readability

Reviewed By: GD06

Differential Revision: D37875461

fbshipit-source-id: 45821616f5edabccaebab06cecb3bae335d6b571