Fix crash during shutdown when profiler was used (#642)
Summary:
The EventProfilerController uses the loggers, so when they get destroyed before the controller you'll get a segfault.
Currently the creation / function calls are in this order:
- profilerMap
- loggers
- makeLoggers

This means the loggers will be destroyed first (reverse order of construction).
This changes puts the profilerMap call to the end destroying that first, then the loggers.

Pull Request resolved: https://github.com/pytorch/kineto/pull/642

Reviewed By: aaronenyeshi

Differential Revision: D38466073

Pulled By: robieta

fbshipit-source-id: b32cea973bac342552a66f9bbbc6ee87a806995d