set device buffer and pool limit
Summary: By default CUPTI uses up to 750MiB worth of CUDA memory and our workloads often OOM. AutoTrace has guard to prevent profiling when there is less than 300MB worth of memory available for tracing. there is no good heuristics on what the right amount of CUPTI buffer is optimal for tracing but we can start with this as default size for now.

Reviewed By: xw285cornell

Differential Revision: D36330887

fbshipit-source-id: 19ab6c93f82ed5a20002bce6113c933d4b46c690