Exit `profilerLoop` after sleep when requested (#641)
Summary:
I see a SEGFAULT on application shutdown as the system is shutting down but the profiler thread is sleeping and then tries to report a last sample but the loggers are already gone.
So check the flag after the sleep to mitigate this.

Actually this is an example of static deinit fiasco. But fixing this is really hard...

Pull Request resolved: https://github.com/pytorch/kineto/pull/641

Reviewed By: slgong-fb

Differential Revision: D38958756

Pulled By: aaronenyeshi

fbshipit-source-id: cf106e951c676479b7fa1012a8e3dc3da11311ce