Add support for reading SLURM job id (#653)
Summary:
SLURM is a popular Job scheduler for open source environments. This PR helps us detect the SLURM Job id in Kineto and this can be used to target this job for on-demand tracing (instead of knowing the pid of a process).
- The job id is read from the process' environment.
- Also updates config loader to use ipcfabric by default if user sets KINETO_USE_DAEMON=1

Note: it only supports numeric job ids right now, and defaults to 0 otherwise.

Pull Request resolved: https://github.com/pytorch/kineto/pull/653

Reviewed By: slgong-fb

Differential Revision: D39369861

Pulled By: briancoutinho

fbshipit-source-id: 605910a061851d7b4406aefcfd648266bf150604