Disable EventProfiler::stop() if it was never started
Summary: Guard the EventProfiler::stop with the same check as EventProfiler::start, so that we are not trying to stop what is not enabled. If we call this, it will create the profilerMap() singleton, which will run into an SIOF bug on the destruction of EventProfilerController and the profilerMap singleton.

Reviewed By: chaekit

Differential Revision: D41097646

Pulled By: aaronenyeshi

fbshipit-source-id: 8937a609fe14830867746231df78d33f4ff46333