Disable EventProfiler::stop() if it was never started (#680)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/680

Guard the EventProfiler::stop with the same check as EventProfiler::start, so that we are not trying to stop what is not enabled. If we call this, it will create the profilerMap() singleton, which will run into an SIOF bug on the destruction of EventProfilerController and the profilerMap singleton.

Test Plan: CI

Reviewed By: chaekit

Differential Revision: D41097646

Pulled By: aaronenyeshi

fbshipit-source-id: e75dd5e23988dccf0d8f48a5bc85c460b92f4638