Remove std::endl which makes json invalid when including logs (#703)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/703

The traces are recording the LOGs for INFO and WARNING, however the strings need to be JSON friendly. In this case, the std::endl will break the json parser. Removing it will enable the trace viewer again.

Test Plan:
This caused several traces to fail to load from this job:
https://www.internalfb.com/intern/zoomer/?profiling_run_fbid=858814338877438&tab=overview

Before this patch:
{F817875595}

After removing the std::endl, I can load the trace:
{F817874957}

Reviewed By: leitian

Differential Revision: D41826592

Pulled By: aaronenyeshi

fbshipit-source-id: 1de9b2054ed7dc9fc54ffa323fed5cbc8167bac7