Sanitize JSON strings for event names
Summary:
This JSON string shows events in the chrome trace. The names for these events can sometimes be file paths, e.g. for python tracing. On windows, file paths contain backslashes, so we need to sanitize these strings.

I also changed the sanitization function so that it always removes backslashes instead of only doing it on windows. That should make testing easier (and we probably don't want backslashes anyway?)

Differential Revision: D45095675

fbshipit-source-id: 08052af2f2c58ee5d2d20d1b6e345871211e5089