Sanitize JSON strings for event names (#754)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/754

This JSON string shows events in the chrome trace. The names for these events can sometimes be file paths, e.g. for python tracing. On windows, file paths contain backslashes, so we need to sanitize these strings.

I also changed the sanitization function so that it always removes backslashes instead of only doing it on windows. That should make testing easier (and we probably don't want backslashes anyway?)

Reviewed By: aaronenyeshi

Differential Revision: D45095675

fbshipit-source-id: 6c528ecf9427d47f4467024a88a9fdae887a55a2