Revert D42825031: Multisect successfully blamed D42825031 for test or build failures
Summary:
This diff is reverting D42825031
D42825031: [wip] kineto add synchronization events in the trace by briancoutinho has been identified to be causing the following test or build failures:

Tests affected:
- [caffe2/caffe2:caffe2_test_gpu - GemmBatchedGPUTrans/GemmBatchedGPUTest.GemmBatchedGPUFloatTest/1](https://www.internalfb.com/intern/test/281475046984558/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2245516
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: aaronenyeshi

Differential Revision: D46679602

fbshipit-source-id: 8376c9e0b3685218fa56801f44825b6af1679756