Update to C++17 and make Config::kControllerIntervalMsecs in… (#791)
Summary:
PyTorch is on C++17 now, so Kineto should be too.

The specific issue was with Config::kControllerIntervalMsecs. This is a static constexpr member - which in pre-C++17 needed a separate namespace definition (i.e. both a declaration `static constexpr chronos::milliseconds kControllerIntervalMsecs` in the class definition and a definition `constexpr chronos::milliseconds Config::kControllerIntervalMsecs`); but after C++17 static constexpr members don't need a separate definition.

Updating to C++17 seems to fix the issue. We could probably remove the `#if __cplusplus < 201703L` handling entirely, but I'm not 100% sure every user of kineto is on C++17

Pull Request resolved: https://github.com/pytorch/kineto/pull/791

Reviewed By: aaronenyeshi

Differential Revision: D47743393

Pulled By: davidberard98

fbshipit-source-id: a38a8f105344808f239ec3d4da6626b3ffc7c356