Omit CUDA sync events on unused CUDA streams (#795)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/795

CUDA WaitforStream sync events can lead to a lot of verbosity in the traces.
It looks like PyTorch cycles through all the streams to add these wait calls but they are no-ops on stream that are not in active use.
 {F1063309394}

This change will simply filter out WaitforStream events whenever a CUDA stream has not been used.
* Track device, stream used in a set, this updated by an GPU kernel or memcpy or memset.
* Filter waitforstream events.

Reviewed By: chaekit, davidberard98

Differential Revision: D47975637

fbshipit-source-id: 26796b6e703f12cdb91bbe99c75db6d89a4c866f