Collect more info on cuduEventRecord for stream wait sync events (#808)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/808

## Overview
To effectively understand synchronization due to CUDA events we also need to track CUDA event record calls and which CUDA event record call the sync occurss on.
Thus when an event synchronization occurs we track
1. We can figure out which cudaEventRecord() this sync took place on
2. Based on that we know which kernel/memcpy launch was recorded by the event, this will typically be the previous kernel/memcpy launch in runtime.

CUPTI' CUDA Event record tells us the correlation ID of the CUDA record [documentation](https://docs.nvidia.com/cupti/annotated.html#structCUpti__ActivityEvent)
```
uint32_t CUpti_ActivityEvent::correlationId [inherited]
The correlation ID of the event. Use of this ID is user-defined, but typically this ID value will equal the correlation ID of the kernel for which the event was gathered.
```

## In this change
* Enable logging cuda Event Record Runtime events (and event sychronize calls too).
* Track the correlation ID for CUPTI Event records as shown above.
* In the CUPTI Stream Wait event we emit the correlation

Reviewed By: anupambhatnagar, davidberard98

Differential Revision: D49694339

fbshipit-source-id: 68b5ca2cc2cdfee6a4b1c90d213813695812e02a