Roctracer: Implement reliable data flushing on stop (#824)
Summary:
When stopping and flushing with libroctracer there is no way to guarantee all completion notifications have been received and logged.   Torch.Profiler calls deviceSynchronize() on stopping the tracer but that ensures nothing.  It does add some delay, reducing the race.

I augmented the tracer stop procedure to enqueue an extra async op and delay until the notification for it is received.

This will fix many "flakey" unit tests that look for kernel executions that occur immediately before the end of a recording.

Pull Request resolved: https://github.com/pytorch/kineto/pull/824

Reviewed By: aaronenyeshi, xuzhao9

Differential Revision: D50841060

Pulled By: briancoutinho

fbshipit-source-id: af736989517a1f958bc657d09efbf3c6d4c897ef