Rework roctracer shutdown flushing (#833)
Summary:
Reworked roctracer flush on shutdown.  Removing a race condition added while removing the original race condition. :)
Previous implementation had a deadlock in the case where the op buffer had filled and flushed on its own immediately before shutdown.

This new approach is simpler but has to continuously log correlation ids of completed async ops.  This is done on the roctracer supplied callback thread so it is not an overhead/performance issue.  Previous attempt was over-optimized at the cost of not working correctly.  Oops.

Pull Request resolved: https://github.com/pytorch/kineto/pull/833

Reviewed By: xuzhao9

Differential Revision: D51473804

Pulled By: aaronenyeshi

fbshipit-source-id: 7a7bea1356aea7e9a719fffe8e50735d072843f6