Fix roctracer flushing issue (#840)
Summary:
Discovered an issue with libroctracer where, immediately after startup, records are not delivered.  Not lost, just not delivered, even on a roctracer_flush_activity() call.

A hipDeviceSynchronize() resolves the issue.  Normally the application, e.g. torch.profiler, would do this synchronize before stopping the profiler.   It looks like this is not always the case though, so we need to inject an additional call.  Without it, stopLogging() may poll forever waiting for data that never comes.

Also, added a timeout to the polling loop.  This will prevent hangs.  Unit tests were the only known examples of the hang.  They should be able to detect the incorrect results if the timeout triggers.  Otherwise "no harm, no foul" applies. :)

Pull Request resolved: https://github.com/pytorch/kineto/pull/840

Reviewed By: xuzhao9

Differential Revision: D51670885

Pulled By: aaronenyeshi

fbshipit-source-id: 6040a0bb40dd1d4f7f8e63bdc8355a6d413119ec