Specify cuContext in range profiler structs (#643)
Summary:

## Context:
I was experiencing some crashes with the range proiler. The backtrace showed ::disable() failing and I found that the `.ctx` member of the configuration struct was not set.

## Changes:
This diff specifies cuContext for all structs that have it as a required member.

This diff also specifies the `.targetNestingLevel` member for the `setConfigParams` struct as it was unset.

Differential Revision: D38435331