Specify cuContext in range profiler structs (#643)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/643

## Context:
I was experiencing some crashes with the range proiler. The backtrace showed ::disable() failing and I found that the `.ctx` member of the configuration struct was not set.

## Changes:
This diff specifies cuContext for all structs that have it as a required member.

This diff also specifies the `.targetNestingLevel` member for the `setConfigParams` struct as it was unset.

Reviewed By: benghaem

Differential Revision: D38435331

fbshipit-source-id: 3a2a297f7ed5665ed3d58df7f48ae8a64e5ac956