Guard stopProfiler callback for EventProfiler only
Summary:
The stopProfiler callback function is only required for EventProfiler to stop it if it on RESOURCE_CONTEXT_DESTROYED. We do not need to register or enable this callback for regular activities tracing.

This is part of our effort to disable callbacks that are not required. See https://partners.nvidia.com/bug/viewbug/4467267

Differential Revision: D53576318

Pulled By: aaronenyeshi