Guard stopProfiler callback for EventProfiler only (#865)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/865

The stopProfiler callback function is only required for EventProfiler to stop it if it on RESOURCE_CONTEXT_DESTROYED. We do not need to register or enable this callback for regular activities tracing.

This is part of our effort to disable callbacks that are not required. See https://partners.nvidia.com/bug/viewbug/4467267

Test Plan: CI

Reviewed By: chaekit

Differential Revision: D53576318

Pulled By: aaronenyeshi

fbshipit-source-id: 1cbdcf15640f580734522e6d70f8e1302298bc5f