Only EventProfiler APIs need to be called under lock (#867)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/867

We don't need to hold the mutex for regular init, only the EventProfiler APIs need to be called under lock.

This helps clean up the logic during init. This is part of the effort to clean up CUPTI callbacks.

Test Plan:
CI

Only two callsites for initMutex, and they are for the start and stop of the EventProfiler only.

Reviewed By: chaekit

Differential Revision: D53576588

Pulled By: aaronenyeshi

fbshipit-source-id: 204e2927deed485087d51544e7fc3c2013cc9dd6