Fix CUPTI timestamp on Windows (#868)
Summary:
I am working on enabling CUPTI on pytorch on Windows https://github.com/pytorch/pytorch/pull/111568 but I had some flaky tests in the CI caused by https://github.com/pytorch/pytorch/blob/main/test/test_foreach.py#L65. After some debugging, I realized that sometimes the tensor_apply_kernel will be `TraceActivity outside of profiling window:`. The issue is caused by the fact that on Windows timestamp from CUPTI is on Windows the monotonic time so we try to sync the system clock with the steady clock.

But after CUDA 11.6 it was added [cuptiActivityRegisterTimestampCallback ](https://forums.developer.nvidia.com/t/cuda-profiler-tools-interface-cupti-for-cuda-toolkit-11-6-is-now-available/200030) for which we register a callback that returns the system time which seems to be the better option.

This branch is used in https://github.com/pytorch/pytorch/pull/111568 which runs the unit tests which pass.
Let me know if there are some additional tests I should do.

Pull Request resolved: https://github.com/pytorch/kineto/pull/868

Reviewed By: xuzhao9

Differential Revision: D54129205

Pulled By: aaronenyeshi

fbshipit-source-id: 032616972d8e3b6cc89c3363689af6658e6ec18b