Roctracer: Fix contention for high QPS cases and add upper limit for buffer size
Summary:
This diff fixes the crashing issue when number of samples is high.
- Adds a mutex for activity_callback and and api_callback when adding rows into buffer
- Adds a thread_local mutex and unordered_map in api_callback due to jemalloc SIGABRT, seems like contention on timestamp.
- Use a shared insert_row_to_buffer() function to lock when storing the row to buffer.

FIXME: We don't currently delete the rows in the buffer, since we need to keep them around until the chrome trace is created. We should fixed this by transferring ownership to CuptiActivityProfiler via Activity Buffer similar to Cupti Activity.

Reviewed By: houseroad, davidberard98

Differential Revision: D55341766

Pulled By: aaronenyeshi