Initialize BackwardNode with start_time, end_time and tid in torch_tb_profiler (#900)
Summary:
Currently when building backward modules in torch_tb_profiler, we initialize `start_time` and `end_time` to `None`, and `tid` to `0` and try to fill these three fields in a bottom–up manner. However, during this process if at least one of the leaf nodes is a `BackwardNode` and the initial values of `start_time` and `end_time` are `None`, we will encounter errors in `self.children.sort(key=lambda x: (x.start_time, -x.end_time))` when calling `fill_stats`. The parent's `tid` can also be set to 0. To fix this bug, this PR assigns a `BackwardNode` with initial values of `start_time`, `end_time`, `tid`. These values can be updated later when traversing the tree backwards.

Pull Request resolved: https://github.com/pytorch/kineto/pull/900

Reviewed By: xuzhao9

Differential Revision: D55923121

Pulled By: aaronenyeshi

fbshipit-source-id: 855a4c4fa5ea31338681925355ceb11fd32ca45a