Fix roctracer crash when number of samples too high (#902)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/902

There is a map access at the bottom of the `api_callback()` function that is storing the correlation id and external correlation id relation. This map is shared between threads, and is causing a SIGSEGV crash during RBTree insert_and_rebalance.

To fix this, we add a mutex for the map. This is likely why the previous thread_local mutex was fixing crashes, due to delayed timing, so we can remove the thread_local mutex now.

Resolves: https://github.com/pytorch/kineto/issues/894

Test Plan: CI and ran on internal high QPS job

Reviewed By: 842974287

Differential Revision: D56017970

Pulled By: aaronenyeshi

fbshipit-source-id: 68dc289740e2998a5b8a17fd0ee54c4a8468098d