Change relative time of 'ts' field from init time to the start of 3 month intervals
Summary:
Use the start of 3 month intervals as the base_time which all the timestamps will be relative to. This is better than using init time of the tracer because traces across ranks are starting at different starting points, and its very hard to correlate them especially when working on PP, FSDP, etc.

There are tools like Chrome Trace Viewer that uses double to represent
each element in the timeline. Double has a 53 bit mantissa to support
up to 2^53 significant digits (up to 9007199254740992). This holds at the
nanosecond level, about 3 months and 12 days. So, let's round base time to
3 months intervals, so we can still collect traces across ranks relative to
each other. A month is 2629746, so for 3 months std::ratio is 7889238.

Reviewed By: sraikund16

Differential Revision: D56642236

Pulled By: aaronenyeshi

fbshipit-source-id: cd66d85181036b75c50c73354215b96d468e9790