Delay logging for stream wait event till end to ensure filtering is (#908)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/908

Fixes https://github.com/pytorch/kineto/issues/904

Today to avoid noisy "even sync" events we do some filtering such that if a kernel is not launched on a stream till now (in trace window) we do not emit these events. But this may be incorrect if a future kernel is present but we just did not see it yet in the events.

Fix: the idea is to not filter the CUDA event synchronizations on the fly. instead defer the logging of some of these events till all other events are processed. We then log these events only if a stream has some activity to avoid noise.

Reviewed By: aaronenyeshi

Differential Revision: D56198193