Add Base Timestamp to Kineto Tracings
Summary: Because we use relative timestamps in Kineto tracings, other plugins that use stitching, like GPUSnoop, have began to fail. In order for events to be stitched properly, they will need the basetime to get a relative timestamp off of. We output this in our chrome trace JSON as "baseTimeNanoseconds"

Differential Revision: D57186263