Add Base Timestamp to Kineto Tracings (#928)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/928

Because we use relative timestamps in Kineto tracings, other plugins that use stitching, like GPUSnoop, have began to fail. In order for events to be stitched properly, they will need the basetime to get a relative timestamp off of. We output this in our chrome trace JSON as "baseTimeNanoseconds"

Reviewed By: aaronenyeshi

Differential Revision: D57186263

fbshipit-source-id: f7628bebda7c40c74a558c125c81aa96312df0a1