Revert D55262038: Multisect successfully blamed "D55262038: Support CUPTI Range Profiler for H100" for one test failure
Summary:
This diff reverts D55262038
D55262038: Support CUPTI Range Profiler for H100 by pls331 causes the following test failure:

Tests affected:
- [cogwheel:cogwheel_adranker_local_compare_test#main](https://www.internalfb.com/intern/test/281475018584027/)

Here's the Multisect link:
https://www.internalfb.com/multisect/5091629
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: aaronenyeshi

Differential Revision: D57370679

fbshipit-source-id: d10c0c84a57d3707ba6f0859756472c9eb7c09e4