Add Join to ManifoldChromeTraceLogger (#938)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/938

Requested in this post:
https://fb.workplace.com/groups/ai.efficiency.tools.users/permalink/1777663889385748/

Manifold upload fail because resources get cleaned by parent thread while manifold is still being uploaded to. In order to fix this, we save any uploading thread as an instance variable. Then, during the destructor, join on this thread if possible.

Currently, the ChromeTraceLogger is declared locally so it will deconstruct once the save() function is exited. This will result in a join() happening right after the thread launch, which negates the use of multithreading. For this reason, we save the ChomeLogger  as an instance variable of the MemLogger. That way, when the CuptiTraceActivity class exits, the MemLogger will exit and then the ChromeTraceLogger will exit thereby calling its destructor (which is much after the save() gets called).

Reviewed By: sanrise

Differential Revision: D57741999

fbshipit-source-id: 15e80130d728ab88cf357f3748e8dc465cdb3778