Lower GPU User Annotation Margin
Summary: Users have reported that some GPU events overlap on only a single side. This is partially because we add a margin to our GPU User Annotations so that they don't have the same exact start and end time as other events if there is only one event in a user annotation. We can lower the margin now that we don't have rounding issues.

Differential Revision: https://internalfb.com/D58254807