Lower GPU User Annotation Margin (#944)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/944

Users have reported that some GPU events overlap on only a single side. This is partially because we add a margin to our GPU User Annotations so that they don't have the same exact start and end time as other events if there is only one event in a user annotation. We can lower the margin now that we don't have rounding issues.

Reviewed By: aaronenyeshi

Differential Revision: D58254807

fbshipit-source-id: 6875fee26f57aba12cde9466d142c9dd62a67146