Add sortIndex to fix ordering of roctracer GPU id tracks (#926)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/926

Although hipGetDeviceProperties shows 8 devices enumerating from 0 to 7, when using roctracer_record_t, the record->device_id enumerates from 2 to 9. This is because roctracer considers id 0 and 1 as CPU sockets.

So we can expect up to 8 CPU sockets and 8 GPUs, up to a total count of 16. Which would mean the enumeration will be 0 - 15. This will help put the GPU tracks in order and at the bottom of the chrome trace.

Original Bug report in ROCm/roctracer: https://github.com/ROCm/roctracer/issues/98. We can just show the device id that is given by roctracer.

Test Plan: CI, added a unit test: JsonGPUIDSortTest, and ran locally.

Reviewed By: chaekit

Differential Revision: D57067065

Pulled By: aaronenyeshi