Add JustKnobs to TSC CUPTI Callback (#947)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/947

Add JustKnobs so that we can turn TSC CUPTI Callback on and off. Since we need to meet open source compliance, we need to first add a default value of true in the Config class. In the FBConfig class we then add a method that checks the JK and overrides the default value of true. We also add a static variable to the CuptiActivityProfiler which will be overriden by the Config value upon initialization. Once this is set, the profiler as well as all headers will use the flag that was set by the JK. Using this flag we can essentially use the TSC timestamp or system clock via killswitch. Since the JK can run asynchronously across many initializations, wrap it in a try/catch.

On top of this change, ApproximateClock is moved from src/ to include/. This is because it is a forward facing header so it makes sense to expose it as a public header. Added also to the bazel file as a public header.

Reviewed By: aaronenyeshi

Differential Revision: D58485370

fbshipit-source-id: f69a331d11940767090718207913ae963516bbb0