Add API for JK Inside Config Validation
Summary: Right now we initialize all the configuration children when config is initialized. This is fine for the most part except for the fact that we set the selected activities only after the config is initialized. This makes our JustKnobs calls always get ignored. We can fix this by calling an API to check the Knobs after the parent is validated (we set activities before validation)

Differential Revision: D58759304