Add API for JK Inside Config Validation (#954)
Summary:
Pull Request resolved: https://github.com/pytorch/kineto/pull/954

Right now we initialize all the configuration children when config is initialized. This is fine for the most part except for the fact that we set the selected activities only after the config is initialized. This makes our JustKnobs calls always get ignored. We can fix this by calling an API to check the Knobs after the parent is validated (we set activities before validation). The API (setActivityDependentConfig) in AbstractConfig will iterate through all child configs and call the child-defined version of setActivityDependentConfig. In FBConfig, will contain the JK reading.

Reviewed By: aaronenyeshi

Differential Revision: D58759304

fbshipit-source-id: ebd0d978627299b5ddecba58fc934e9342b31da3