Add Dist Info for On-Demand NCCL Traces
Summary:
Currently, auto-trace uses pytorches c10d.distributed library to set up the "distributedInfo" field in the chrome tracing. This makes sense as all the process groups are initialized and maintained in python. Because of this, however, on-demand is unable to retrieve this information directly as it has no access to the python side of a graph. However, we can glean the process group information based on the information that is added in to the dispatcher.

In this diff, we add a rank parameter to the nccl information during collection. We then keep a global to enumerate all the nccl information via the args in the collectives as we print them out to the trace. At the end of the trace we output the distributed information. In order to prevent double counting for auto-trace, we skip if there was previously distributedInfo ouputted to the trace.

NOTE: This diff contains kineto + profiler changes so it will need to be broken up into two seperate diffs

Differential Revision: D58736045